requirements.txt
beautifulsoup4 == 4.12.3
requests == 2.32.3
pandas == 2.1.4
openpyxl == 3.1.5
python-dotenv==1.0.1